requirements.txt
f1-2020-telemetry==0.2.1
Telemetry-F1-2021
signalfx
PyYAML
